\name{EMdir-package}
\alias{EMdir-package}
\alias{EMdir}
\docType{package}
\title{
\packageTitle{EMdir}
}
\description{
\packageDescription{EMdir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMdir}
\packageIndices{EMdir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMdir}

Maintainer: \packageMaintainer{EMdir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
